API

Topology

Create and apply topology

Mango.complete_topology — Function

complete_topology(number_of_nodes)

Create a fully-connected topology.

Mango.star_topology — Function

star_topology(number_of_nodes)

Create a star topology.

Mango.cycle_topology — Function

cycle_topology(number_of_nodes)

Create a cycle topology.

Mango.graph_topology — Function

graph_topology(graph)

Create a topology based on a Graphs.jl (abstract) graph.

Mango.create_topology — Function

create_topology(create_runnable)::Topology

Create a topology using the create_runnable function which is a one-argument function with an initially empty topology as argument.

Example

topology = create_topology() do topology
    agent = register(container, TopologyAgent())
    agent2 = register(container, TopologyAgent())
    agent3 = register(container, TopologyAgent())
    n1 = add_node!(topology, agent)
    n2 = add_node!(topology, agent2)
    n3 = add_node!(topology, agent3)
    add_edge!(topology, n1, n2)
    add_edge!(topology, n1, n3)
end

Mango.per_node — Function

per_node(assign_runnable, topology)

Loops over the nodes of the topology, calls assign_runnable on every node to enable the caller to populate the node. After the loop finished the neighborhoods are created and injected into the agent.

Example

per_node(topology) do node
    add!(node, register(container, TopologyAgent()))
end

Modify data structures

Mango.add! — Function

add!(node, agent::Agent...)

Add an agents to the node.

Mango.add_node! — Function

add_node!(topology, agents::Agent...)::Int

Add a node to the topology with a list (or a single) of agents attached.

Graphs.SimpleGraphs.add_edge! — Function

add_edge!(topology, node_id_from, node_id_to, directed=false)

Add an edge to the topology from node_id_from to node_id_to. If directed is true a directed edge is added, otherwise an undirected edge is added.

Use the topology

Mango.topology_neighbors — Function

topology_neighbors(agent)

Retrieve the neighbors of the agent, represented by their addresses. These vaues will be updated when a topology is applied using per_node or create_topology.

Autodocs

Mango.Address — Type

Supertype of all address types

Mango.Agent — Type

Base-type for all agents in mango. Generally exists for type-safety and default implementations across all agents.

Mango.address — Method

Used internally by the role to get the AgentAddress

Mango.aid — Method

Used internally by the role to get the AID

Mango.forward_to — Method

API Definition for forwarding messages to other agents.

Mango.reply_to — Method

API Definition for directly replying to a message

Mango.send_and_handle_answer — Method

API Definition for sending tracked message with required response handler

Mango.send_message — Method

API Definition for the role context

Mango.AgentContext — Type

Context of the agent. Represents the environment for the specific agent. Therefore it includes a connection to the container, including all functions used for interacting with the environment for the agent.

Mango.AgentRoleHandler — Type

Internal data regarding the roles.

Base.schedule — Method

Delegates to the scheduler Scheduler

Mango.add — Method

Add a role to the agent. This will add the role to the internal RoleHandler of the agent and it will bind the RoleContext to the role, which enables the role to interact with its environment.

Mango.add_forwarding_rule — Method

Add a rule for message forwarding.

After calling the agent will auto-forward every message coming from from_addr to to_address. If forwardreplies is set, all replies from `toaddressare forwarded back tofrom_addr`.

Mango.add_service! — Method

add_service!(agent, service)

Add a service to the agent. Every service can exists exactly one time (stored by type).

Mango.address — Method

Shorter Alias

Mango.aid — Method

Returns the agent id of the agent.

Mango.delete_forwarding_rule — Method

Delete an added forwarding rule. If to_address is not set, all rules are removed matching from_addr. If it set, both addresses need to match.

Mango.forward_to — Method

forward_to(agent, content, forward_to_address, received_meta; kwargs...)

Forward the message to a specific agent using the metadata received on handling the message. This method essentially simply calls send_message on the input given, but also adds and fills the correct metadata fields to mark the message as forwarded.

For this the following meta data is set. 'forwarded=true', 'forwardedfromaddress=address of the original sender', 'forwardedfromid=id of the original sender'

Mango.handle_message — Method

Defines a function for an agent, which will be called when a message is dispatched to the agent. This methods will be called with any arriving message (according to the multiple dispatch of julia).

Mango.on_ready — Method

Lifecycle Hook-in function called when the agent system as a whole is ready, the hook-in has to be manually activated using notify_ready(container::Container)

Mango.on_start — Method

Lifecycle Hook-in function called when the container of the agent has been started, depending on the container type it may not be called (if there is no start at all, f.e. the simulation container)

Mango.reply_to — Method

Convenience method to reply to a received message using the meta the agent received. This reduces the regular sendmessage as response `sendmessage(agent, "Pong", AgentAddress(aid=meta["senderid"], address=meta["senderaddr"]))toreply_to(agent, "Pong", meta)`

Furthermore it guarantees that agent address (including the tracking id, which is part of the address!) is correctly passed to the mango container.

Mango.roles — Method

Return all roles of the given agent

Mango.send_and_handle_answer — Method

Convenience method for sending tracked messages with response handler to the answer.

Sends a tracked message with a required response_handler to enable to use the syntax

send_and_handle_answer(...) do agent, message, meta
	# handle the answer
end

Mango.send_message — Method

Send a message using the context to the agent with the receiver id receiver_id at the address receiver_addr. This method will always set a sender_id. Additionally, further keyword arguments can be defines to fill the internal meta data of the message.

Mango.send_tracked_message — Method

Send a message using the context to the agent with the receiver id receiver_id at the address receiver_addr. This method will always set a sender_id. Additionally, further keyword arguments can be defines to fill the internal meta data of the message.

Furthermore, message sent with this method will be wrapped in a data object which annotates the message with a transactional id, to be able to track this specific agent discussion. For this it is possible to define a response_handler, to which a functin can be assigned, which handles the answer to this message call. To continue the conversation the transaction id has to be tr by kwargs in the response handler

Mango.service_of_type — Method

service_of_type(agent, type::Type{T}, default=nothing)::Union{T,Nothing} where {T}

Return the current agent service of the type type.

If a default is set, this default service will be added to the agent as service of te type `type. The function is especially useful if you want to extend the functionality of the agent without having to change the internals of the agent, as this functions enables the user to add arbitrary data to the agent on which functions can be defined.

Mango.services — Method

services(agent)::Dict{DataType,Any}

Return a list of services, which were added to the agent.

Mango.shutdown — Method

Will be called on shutdown of the container, in which the agent is living

Mango.stop_all_tasks — Method

Delegates to the scheduler Scheduler

Mango.stop_and_wait_for_all_tasks — Method

Delegates to the scheduler Scheduler

Mango.stop_task — Method

Delegates to the scheduler Scheduler

Mango.wait_for_all_tasks — Method

Delegates to the scheduler Scheduler

Mango.@agent — Macro

Macro for defining an agent struct. Expects a struct definition as argument.

The macro does 3 things:

It adds all baseline fields, defined in AGENTBASELINEFIELDS (the agent context context, the role handler role_handler, and the aid)

It adds the supertype Agent to the given struct.

It defines a default constructor, which assigns all baseline fields to predefined default values. As a result you can (and should) create an agent using only the exclusive fields.

For example the usage could like this.

@agent struct MyAgent
	my_own_field::String
end

# results in

mutable struct MyAgent <: Agent
	# baseline fields...
	my_own_field::String
end
MyAgent(my_own_field) = MyAgent(baseline fields defaults..., my_own_field)

# so youl would construct your agent like this

my_agent = MyAgent("own value")