API

Express

This part contains basic convenience functions for creating and running Mango.jl simulations.

Mango.activate — Method

activate(runnable, container_list)

Actvate the container(s), which includes starting the containers and shutting them down after the runnable has been executed.

In most cases the runnable will execute code, which starts some process (e.g. some distributed negotiation) in the system to define the objective of the agent system.

Generally this function is a convenience function and is equivalent to starting all containers in the list, executing the code represented by runnable and shuting down the container again. Further, this function will handle errors occuring while running the runnable and ensure the containers are shutting down.

Examples

activate(your_containers) do 
   # Send the first message to start the system
   send_message(defined_agent, "Starting somethin", address(other_defined_agent))

   # wait some time
   wait(some_stopping_condition)
end

Mango.add_agent_composed_of — Method

add_agent_composed_of(container, roles...)

Create an agent which is composed of the given roles roles... and register the agent to the container.

The agent struct used is an empty struct, it is only used as a container for the roles. The created agent will be returned by the function.

Examples

agent = add_agent_composed_of(your_container, RoleA(), RoleB(), RoleC())

Mango.create_mqtt_container — Method

create_mqtt_container(host, port, client_id; codec=nothing)

Create a container using an MQTT protocol.

The host is expected to be the IP-adress of the broker, port is the port of the broker, the client_id is the id of the client which will be created and connected to the broker. Optionally you can also provide a codec as tuple of functions (first encode, second decode, see encode and decode).

Examples

agent = create_mqtt_container("127.0.0.1", 5555, "MyClient")

Mango.create_tcp_container — Method

create_tcp_container(host, port; codec=nothing)

Create a container using an TCP protocol.

The host is expected to be the IP-adress to bind on, port is the port to bind on. Optionally you can also provide a codec as tuple of functions (first encode, second decode, see encode and decode).

Examples

agent = create_mqtt_container("127.0.0.1", 5555, "MyClient")

Agent and Roles

Here, the API for the agent structs created with @agent/@role is listed.

Mango.address — Method

address(agent)

Return the agent address of the agent as AgentAddress or MQTTAddress depending on the protocol.

Mango.aid — Method

Used internally by the role to get the AID

Mango.forward_to — Method

forward_to(agent, content, forward_to_address, received_meta; kwargs...)

Forward the message to a specific agent using the metadata received on handling the message. This method essentially simply calls send_message on the input given, but also adds and fills the correct metadata fields to mark the message as forwarded.

For this the following meta data is set. 'forwarded=true', 'forwardedfromaddress=address of the original sender', 'forwardedfromid=id of the original sender'

Mango.reply_to — Method

reply_to(agent, content, received_meta)

Convenience method to reply to a received message using the meta the agent received. This reduces the regular sendmessage as response `sendmessage(agent, "Pong", AgentAddress(aid=meta["senderid"], address=meta["senderaddr"]))toreply_to(agent, "Pong", meta)`

Furthermore it guarantees that agent address (including the tracking id, which is part of the address!) is correctly passed to the mango container.

Mango.send_and_handle_answer — Method

send_and_handle_answer(
    response_handler::Function(agent, message, meta)::nothing,
    agent::AgentInterface,
    content::Any,
    agent_address::Address;
    calling_object::Any=nothing,
    kwargs...)

Convenience method for sending tracked messages with response handler to the answer.

Sends a tracked message with a required response_handler to enable to use the syntax

send_and_handle_answer(...) do agent, message, meta
	# handle the answer
end

Mango.send_message — Method

send_message(agent, content::Any, agent_address::Address; kwargs...,)

Send a message with the content content to the agent represented by agent_address.

This method will always set a sender_id. Additionally, further keyword arguments can be defined to fill the internal meta data of the message.

Mango.send_tracked_message — Method

send_tracked_message(agent, content, agent_address; 
    response_handler::Function(agent, message, meta)::nothing,
    calling_object::Any=nothing,
    kwargs...,
)

Send a message with the content content to the agent represented by agent_address. This function will set a generated tracking_id to the address, which allows the identification of the dialog.

It is possible to define a response_handler, to which a function can be assigned, which handles the answer to this message call. Note that the resonding agent needs to use the same tracking id in the response, ideally reply_to is used to achieve this automatically.

This method will always set a sender_id. Additionally, further keyword arguments can be defines to fill the internal meta data of the message.

Mango.Address — Type

Supertype of all address types

Mango.Agent — Type

Base-type for all agents in mango. Generally exists for type-safety and default implementations across all agents.

Mango.@agent — Macro

Macro for defining an agent struct. Expects a struct definition as argument.

The macro does 3 things:

It adds all baseline fields, defined in AGENTBASELINEFIELDS (the agent context context, the role handler role_handler, and the aid)

It adds the supertype Agent to the given struct.

It defines a default constructor, which assigns all baseline fields to predefined default values. As a result you can (and should) create an agent using only the exclusive fields.

Example

For example the usage could like this.

@agent struct MyAgent
	my_own_field::String
end

# results in

mutable struct MyAgent <: Agent
	# baseline fields...
	my_own_field::String
end
MyAgent(my_own_field) = MyAgent(baseline fields defaults..., my_own_field)

# so youl would construct your agent like this

my_agent = MyAgent("own value")

Base.schedule — Method

schedule(f::Function, agent::Agent, data::TaskData)

Delegates to the scheduler Scheduler

Mango.add — Method

add(agent::Agent, role::Role)

Add a role to the agent. This will add the role to the internal RoleHandler of the agent and it will bind the RoleContext to the role, which enables the role to interact with its environment.

Mango.add_forwarding_rule — Method

add_forwarding_rule(agent, from_addr::AgentAddress, to_address::AgentAddress, forward_replies::Bool)

Add a rule for message forwarding.

After calling the agent will auto-forward every message coming from from_addr to to_address. If forwardreplies is set, all replies from `toaddressare forwarded back tofrom_addr`.

Mango.add_service! — Method

add_service!(agent, service)

Add a service to the agent. Every service can exists exactly one time (stored by type).

Mango.delete_forwarding_rule — Method

delete_forwarding_rule(agent, from_addr::AgentAddress, to_address::Union{Nothing,AgentAddress})

Delete an added forwarding rule. If to_address is not set, all rules are removed matching from_addr. If it set, both addresses need to match.

Mango.handle_message — Method

handle_message(agent::Agent, message::Any, meta::Any)

Defines a function for an agent, which will be called when a message is dispatched to the agent. This methods will be called with any arriving message (according to the multiple dispatch of julia).

Mango.on_ready — Method

on_ready(agent::Agent)

Lifecycle Hook-in function called when the agent system as a whole is ready, the hook-in has to be manually activated using notify_ready(container::Container)

Mango.on_start — Method

on_start(agent::Agent)

Lifecycle Hook-in function called when the container of the agent has been started, depending on the container type it may not be called (if there is no start at all, f.e. the simulation container)

Mango.roles — Method

roles(agent)

Return all roles of the given agent

Mango.service_of_type — Method

service_of_type(agent, type::Type{T}, default=nothing)::Union{T,Nothing} where {T}

Return the current agent service of the type type.

If a default is set, this default service will be added to the agent as service of te type `type. The function is especially useful if you want to extend the functionality of the agent without having to change the internals of the agent, as this functions enables the user to add arbitrary data to the agent on which functions can be defined.

Mango.services — Method

services(agent)::Dict{DataType,Any}

Return a list of services, which were added to the agent.

Mango.shutdown — Method

shutdown(agent)

Will be called on shutdown of the container, in which the agent is living

Mango.stop_all_tasks — Method

stop_all_tasks(agent::Agent)

Delegates to the scheduler Scheduler

Mango.stop_and_wait_for_all_tasks — Method

stop_and_wait_for_all_tasks(agent::Agent)

Delegates to the scheduler Scheduler

Mango.stop_task — Method

stop_task(agent::Agent, t::Task)

Delegates to the scheduler Scheduler

Mango.wait_for_all_tasks — Method

wait_for_all_tasks(agent::Agent)

Delegates to the scheduler Scheduler

Mango.AgentContext — Type

Context of the agent. Represents the environment for the specific agent. Therefore it includes a connection to the container, including all functions used for interacting with the environment for the agent.

Mango.AgentRoleHandler — Type

Internal data regarding the roles.

Mango.@role — Macro

Macro for defining a role struct. Expects a struct definition as argument.

The macro does 3 things:

It adds all baseline fields, defined in ROLEBASELINEFIELDS (the role context)

It adds the supertype Role to the given struct.

It defines a default constructor, which assigns all baseline fields to predefined default values. As a result you can (and should) create a role using only the exclusive fields.

For example the usage could like this.

@role struct MyRole
	my_own_field::String
end

# results in

mutable struct MyRole <: Agent
	# baseline fields...
	my_own_field::String
end
MyRole(my_own_field) = MyRole(baseline fields defaults..., my_own_field)

# so youl would construct your role like this

my_roel = MyRole("own value")

Mango.@shared — Macro

Mark the field as shared across roles, this will implicitly

Base.schedule — Method

Delegates to the scheduler Scheduler

Mango.address — Method

Get AgentAddress of the parent agent

Mango.aid — Method

Get AID of the parent agent

Mango.bind_context — Method

Internal function, used to initialize to role for a specified agent

Mango.emit_event — Method

Emit an event to their subscriber

Mango.get_model — Method

Get a shared model from the pool. If the model does not exist yet, it will be created. Only types with default constructor are allowed!

Mango.handle_event — Method

Default function for arriving events, which get dispatched to the role.

Mango.handle_message — Method

Default function for arriving messages, which get dispatched to the role. This function will be called for every message arriving at the agent of the role.

Mango.on_ready — Method

Lifecycle Hook-in function called when the agent system as a whole is ready, the hook-in has to be manually activated using notify_ready(container::Container)

Mango.on_start — Method

Lifecycle Hook-in function called when the container of the agent has been started, depending on the container type it may not be called (if there is no start at all, f.e. the simulation container)

Mango.setup — Method

Hook-in function to setup the role, after it has been added to its agent.

Mango.shutdown — Method

Hook-in function, which will be called on shutdown of the roles agent.

Mango.subscribe_event — Method

Subscribe to specific types of events.

Mango.subscribe_event — Method

Subscribe to specific types of events.

Mango.subscribe_message — Method

Subscribe a message handler function (it need to have the signature (role, message, meta)) to the message dispatching. This handler function will be called everytime the given condition function (message, meta -> boolean) evaluates to true when a message arrives at the roles agent.

Mango.subscribe_send — Method

Subscribe a sendmessage hook in function (signature, (role, content, receiverid, receiver_addr; kwargs...)) to the message sending. The hook in function will be called every time a message is sent by the agent.

Mango.Role — Type

Defines the type Role, which is the common base types for all roles in mango.

A Role is a bundled behavivor of an agent, which shall fulfill exactly one responsibility of an agent - the role. Technically speaking roles are the way to implement the composition pattern for agents, and to introduce modular archetypes, which shall be reused in different contexts.

Mango.RoleContext — Type

The RoleContext connects the role with its environment, which is mostly its agents. This is abstracted using the AgentInterface.

Container

This part contains the API related to the container construction, access and management.

Mango.AgentAddress — Type

Default AgentAddress base type, where the agent identifier is based on the container created agent id (aid). Used with the TCP protocol.

Mango.ContainerInterface — Type

Supertype of every container implementation. This acts as an interface to be used by the agents in their contexts.

Mango.MQTTAddress — Type

Connection information for an MQTT topic on a given broker. Used with the MQTT protocol.

Mango.protocol_addr — Method

Used by the agent to get the protocol addr part

Mango.send_message — Function

Send a message message using the given containercontainer` to the given address. Additionally, further keyword arguments can be defines to fill the internal meta data of the message.

This only defines the function API, the actual implementation is done in the core container module.

Mango.Container — Type

The default container struct, representing the container as actor. The container is implemented by composition. This means the container consists of different implementations of base types, which define the behavior of the container itself. That being said, the same container generally able to send messages via different protocols using different codecs.

Mango.notify_ready — Method

Mark the agent system as ready, needs to be detected and called manually!

Mango.protocol_addr — Method

Get protocol addr part

Mango.register — Function

Register an agent given the target container container. While registering an aid will be generated and assigned to the agent.

This function will add the agent to the internal list of the container and will from then on be controlled by the container regarding the messaging activities. That means the container acts as the gateway of the agent defining its possible way to retrieve messages.

Args

suggested_aid: you can provide an aid yourself. The container will always use that aid if possible

Returns

The actually used aid will be returned.

Mango.send_message — Function

Send a message message with using the given container container to the agent with the receiver id receiver_id. The receivers address is used by the chosen protocol to appropriatley route the message to external participants. To specifiy further meta data of the message kwargs should be used.

Returns

True if the message has been sent successfully, false otherwise.

Mango.send_message — Method

Send message version for MQTT topics. Note that there is no local message forwarding here because messages always get pushed to a broker and are not directly addressed to an agennt.

Mango.shutdown — Method

Shut down the container. It is always necessary to call it for freeing bound resources

Mango.start — Method

Starts the container and initialized all its components. After the call the container start to act as the communication layer.

Base.close — Method

close(protocol::MQTTProtocol)

Disconnect the client from the broker and stop the message loop.

Mango.id — Method

id(protocol::MQTTProtocol)

Return the name the client is registered with at its broker.

Mango.init — Method

init(protocol::MQTTProtocol, stop_check::Function, data_handler::Function)

Initialize the Mosquitto looping task for the provided protocol and forward incoming messages to data_handler.

Mango.notify_register — Method

notify_register(protocol::MQTTProtocol, aid::String; kwargs...)

Notify the protocol MQTT client that a new agent with aid has been registered.

Registration expects a kwarg topics to subscribe the agent to. If any topic is not yet subscribed by the client subscribe is called for it.

Mango.parse_id — Method

parse_id(_::MQTTProtocol, id::Any)::String

Return the id of the protocol as string (for compliance with container core).

Mango.send — Method

send(protocol::MQTTProtocol, destination::String, message::Any)

Send a message to topic destination on the clients MQTT broker.

Returns

Return value and message id from MQTT library.

Mango.subscribe — Method

subscribe(protocol::MQTTProtocol, topic::String; qos::Int=1)

Subscribe the MQTT client of the protocol to topic with the given qos setting.

Returns

The Mosquitto error code and the message id.

Mango.Protocol — Type

Type for all implementations of protocols, acts like an interface. A protocol defines the way message are processed and especially sent and received to an other peer. F.E. A protocol could be to send messages using a plain TCP connection, which would indicate that an internet address (host + port) is required for the communication.

The parameterized type T indicates type, which defines the address data of the receiver and sender.

Every protocol has to define two methods.

send: defines the behavior of the protocol when an agents sends a messages

init: defines the necessary steps to initialize (especially) the receiver loop and

 therefore accepts a stop check and a data handler function to indicate when the receiver should stop, 
	 respectively how to dispatch the received message to the correct agent

Mango.id — Method

Return the external identifier associated with the protocol (e.g. it could be the host+port, dns name, ...)

Mango.init — Method

Initialized the protocols internal loops (if exists). In most implementation this would mean that the receiver loop is started. To handle received messages the data_handler function can be passed (msg, sender) -> your handling code.

To control the lifetime of the loops a stop_check should be passed (() -> boolean). If the stop check is true the loops will terminate. The exact behavior depends on the implementation though.

Mango.notify_register — Method

Protocol specific updates called when a new agent is registered.

Mango.parse_id — Method

Parse different types to the correct type (if required). Should be implemented if the id type is not trivial.

Mango.send — Method

Send the message message to the agent known by the adress destination. How the message is exactly handled is determined by the protocol invoked.

The type of the destination has to match with the protocol.

Returns

The function returns a boolean indicating whether the message was successfull sent

Mango.protocol_addr — Method

Get protocol addr part

Mango.register — Function

Register an agent given the target container container. While registering an aid will be generated and assigned to the agent.

This function will add the agent to the internal list of the container and will from then on be controlled by the container regarding the messaging activities. That means the container acts as the gateway of the agent defining its possible way to retrieve messages.

Args

suggested_aid: you can provide an aid yourself. The container will always use that aid if possible

Returns

The actually used aid will be returned.

Mango.send_message — Function

Send a message message with using the given container container to the agent with the receiver id receiver_id. The receivers address is used by the chosen protocol to appropriatley route the message to external participants. To specifiy further meta data of the message kwargs should be used.

Returns

True if the message has been sent successfully, false otherwise.

Mango.shutdown — Method

Shut down the container. It is always necessary to call it for freeing bound resources

Base.close — Method

Release all tcp resources

Mango.init — Method

Initialized the tcp protocol. This starts the receiver and stop loop. The receiver loop will call the data_handler with every incoming message. Further it provides as sender adress a InetAddr object.

Mango.send — Method

Send a message message over plain TCP using destination as destination address. The message has to be provided as a form, which is writeable to an arbitrary IO-Stream.

Returns

Return true if successfull.

Simulation

In the following the APIs regarding the simulation container are listed.

Scheduling

In the following the APIs for scheduling TaskData is listed.

Topology

In the following the APIs for creating, aplying and using topologies is listed.

Graphs.SimpleGraphs.add_edge! — Method

add_edge!(topology, node_id_from, node_id_to, directed=false)

Add an edge to the topology from node_id_from to node_id_to. If directed is true a directed edge is added, otherwise an undirected edge is added.

Mango.add! — Method

add!(node, agent::Agent...)

Add an agents to the node.

Mango.add_node! — Method

add_node!(topology, agents::Agent...)::Int

Add a node to the topology with a list (or a single) of agents attached.

Mango.complete_topology — Method

complete_topology(number_of_nodes)

Create a fully-connected topology.

Mango.create_topology — Method

create_topology(create_runnable)::Topology

Create a topology using the create_runnable function which is a one-argument function with an initially empty topology as argument.

Example

topology = create_topology() do topology
    agent = register(container, TopologyAgent())
    agent2 = register(container, TopologyAgent())
    agent3 = register(container, TopologyAgent())
    n1 = add_node!(topology, agent)
    n2 = add_node!(topology, agent2)
    n3 = add_node!(topology, agent3)
    add_edge!(topology, n1, n2)
    add_edge!(topology, n1, n3)
end

Mango.cycle_topology — Method

cycle_topology(number_of_nodes)

Create a cycle topology.

Mango.graph_topology — Method

graph_topology(graph)

Create a topology based on a Graphs.jl (abstract) graph.

Mango.per_node — Method

per_node(assign_runnable, topology)

Loops over the nodes of the topology, calls assign_runnable on every node to enable the caller to populate the node. After the loop finished the neighborhoods are created and injected into the agent.

Example

per_node(topology) do node
    add!(node, register(container, TopologyAgent()))
end

Mango.star_topology — Method

star_topology(number_of_nodes)

Create a star topology.

Mango.topology_neighbors — Method

topology_neighbors(agent)

Retrieve the neighbors of the agent, represented by their addresses. These vaues will be updated when a topology is applied using per_node or create_topology.

Encoding/Decoding

In the following the built-in functions for encoding and decoding messages are listed.

Mango.decode — Function

decode(buf::Vector{UInt8}, t::Type=OrderedDict{String, Any})

Decode the data in buf into an object of type t using LightBSON.bson_read.

Examples

julia> data = OrderedDict(["test" => 10])
OrderedDict{String, Int64} with 1 entry:
  "test" => 10

julia> decode(encode(data))
OrderedDict{String, Any} with 1 entry:
  "test" => 10

Mango.encode — Method

encode(data::OrderedDict{String,Any})

Encode data into a UInt8 buffer using LightBSON.

Examples

julia> data = OrderedDict(["test" => 10])
OrderedDict{String, Int64} with 1 entry:
  "test" => 10

julia> encode(data)
19-element Vector{UInt8}:
 0x13
 0x00
 0x00
 0x00
 0x12
    ⋮
 0x00
 0x00
 0x00
 0x00
 0x00