API

Topology

Create and apply topology

Mango.complete_topology — Function

complete_topology(number_of_nodes)

Create a fully-connected topology.

Mango.star_topology — Function

star_topology(number_of_nodes)

Create a star topology.

Mango.cycle_topology — Function

cycle_topology(number_of_nodes)

Create a cycle topology.

Mango.graph_topology — Function

graph_topology(graph)

Create a topology based on a Graphs.jl (abstract) graph.

Mango.create_topology — Function

create_topology(create_runnable)::Topology

Create a topology using the create_runnable function which is a one-argument function with an initially empty topology as argument.

Example

topology = create_topology() do topology
    agent = register(container, TopologyAgent())
    agent2 = register(container, TopologyAgent())
    agent3 = register(container, TopologyAgent())
    n1 = add_node!(topology, agent)
    n2 = add_node!(topology, agent2)
    n3 = add_node!(topology, agent3)
    add_edge!(topology, n1, n2)
    add_edge!(topology, n1, n3)
end

Mango.per_node — Function

per_node(assign_runnable, topology)

Loops over the nodes of the topology, calls assign_runnable on every node to enable the caller to populate the node. After the loop finished the neighborhoods are created and injected into the agent.

Example

per_node(topology) do node
    add!(node, register(container, TopologyAgent()))
end

Modify data structures

Mango.add! — Function

add!(node, agent::Agent...)

Add an agents to the node.

Mango.add_node! — Function

add_node!(topology, agents::Agent...)::Int

Add a node to the topology with a list (or a single) of agents attached.

Graphs.SimpleGraphs.add_edge! — Function

add_edge!(topology, node_id_from, node_id_to, directed=false)

Add an edge to the topology from node_id_from to node_id_to. If directed is true a directed edge is added, otherwise an undirected edge is added.

Use the topology

Mango.topology_neighbors — Function

topology_neighbors(agent)

Retrieve the neighbors of the agent, represented by their addresses. These vaues will be updated when a topology is applied using per_node or create_topology.